refactor: move install prompt to a separate component
